requirements.txt
python-dotenv
google-generativeai
numpy<=2.1.1,>=1.23.0
langchain
langchain-community
langchain-core
langchain-google-genai
google-ai-generativelanguage
SpeechRecognition
pyaudio
pyttsx3
sounddevice
scipy